Slim down Docker build
Move all of the settings which change very infrequently up higher in the
Dockerfile to avoid recreating as many layers as possible. We want to
exploit the Docker build cache.

Add a .dockerignore file so that we send as little information as
possible to the Docker daemon. It also shrinks the amount of data in the
final image.

The Django static files are now collected at container startup. This is
an idempotent operation, and so can be safely run at each container
start. It shrinks the amount of data in the final image, and removes a
layer as well.